+++ QuickStartTemplate/projects/QuickStartTemplate-contracts/nft_mint_server/api/index.js
@@ -1,0 +1,3 @@
+// Vercel serverless entry – export the Express app
+import app from '../app.js'
+export default app
