+++ QuickStartTemplate/projects/QuickStartTemplate-contracts/nft_mint_server/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "nft_mint_server",
+  "private": true,
+  "type": "module",
+  "engines": { "node": ">=18 <21" },
+  "dependencies": {
+    "@pinata/sdk": "^2.1.0",
+    "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "multer": "^2.0.2"
+  },
+  "devDependencies": {}
+}
