+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/package.json
@@ -45,8 +45,8 @@
     "notistack": "^3.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tslib": "^2.6.2",
-    "react-icons": "4.12.0"
+    "react-icons": "4.12.0",
+    "tslib": "^2.6.2"
   },
   "scripts": {
     "generate:app-clients": "algokit project link --all",
